requirements.txt
Flask==2.0.1
python-pptx==0.6.21
Werkzeug==2.0.1
gunicorn==20.1.0
